Rename gestalt-server to gestaltd and remove subcommands
Rename the server binary from gestalt-server to gestaltd, following the
daemon naming convention (mongod, httpd, etc.). Remove the subcommand
dispatch (serve, dev, config) in favor of a single daemon that starts
when run. Config validation is now a --check flag, and the dev subcommand
is removed since web/dev.sh already handles multi-process orchestration.
