+++ .github/workflows/publish-gestaltd-image.yml
@@ -1,4 +1,4 @@
-name: Publish Gestalt Server Image
+name: Publish Gestalt Image
 
 on:
   push:
