+++ Dockerfile
@@ -3,10 +3,10 @@ WORKDIR /build
 COPY go.mod go.sum ./
 RUN go mod download
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -o /gestalt-server ./cmd/gestalt-server
+RUN CGO_ENABLED=0 GOOS=linux go build -o /gestaltd ./cmd/gestaltd
 
 FROM gcr.io/distroless/static-debian12
-COPY --from=builder /gestalt-server /gestalt-server
+COPY --from=builder /gestaltd /gestaltd
 EXPOSE 8080
-ENTRYPOINT ["/gestalt-server"]
+ENTRYPOINT ["/gestaltd"]
 CMD ["--config", "/etc/gestalt/config.yaml"]
