--- cmd/gestalt-server/main_test.go
@@ -1,17 +1,0 @@
-package main
-
-import (
-	"testing"
-)
-
-func TestRun_UnknownCommand(t *testing.T) {
-	t.Parallel()
-
-	err := run([]string{"bogus"})
-	if err == nil {
-		t.Fatal("expected error for unknown command")
-	}
-	if err.Error() != "unknown command: bogus" {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
