+++ docs/pages/contributing.mdx
@@ -4,7 +4,7 @@
 
 | Directory | Purpose |
 |---|---|
-| `cmd/gestalt-server` | Main Go binary. Registers all plugins and starts the server. |
+| `cmd/gestaltd` | Main Go binary. Registers all plugins and starts the server. |
 | `core` | Public interfaces: auth providers, datastores, secret managers, providers, runtimes, bindings, hooks. |
 | `internal` | Runtime packages: config loading, HTTP routing, middleware, invocation, token management, bootstrap. |
 | `plugins` | Implementations of core interfaces: auth, datastores, secrets, integrations, bindings, runtimes. |
@@ -17,7 +17,7 @@
 
 1. Implement the interface from `core` (e.g. `core.AuthProvider`, `core.Datastore`, `core.SecretManager`).
 2. Expose a factory function that accepts the provider config and returns an initialized instance.
-3. Register the factory in `cmd/gestalt-server/factories.go` so the runtime resolves it by name from config.
+3. Register the factory in `cmd/gestaltd/factories.go` so the runtime resolves it by name from config.
 4. Add tests that exercise the plugin through the same interface the runtime uses.
 
 ## Add integration-specific behavior
@@ -26,7 +26,7 @@ Integration hooks live under `plugins/integrations`. Each package exports named 
 
 1. Create a package under `plugins/integrations/<name>`.
 2. Implement the relevant hook signatures (`response_check`, `token_parser`, or `request_mutator`).
-3. Register them by name in `cmd/gestalt-server/factories.go`.
+3. Register them by name in `cmd/gestaltd/factories.go`.
 
 ## Development workflow
 
@@ -36,10 +36,10 @@ cd gestalt
 go test ./...
 
 # Validate config
-gestalt-server config check -config gestalt.dev.yaml
+gestaltd --check --config gestalt.dev.yaml
 
 # Run locally
-gestalt-server dev -config gestalt.dev.yaml
+gestaltd --config gestalt.dev.yaml
 
 # Web UI
 cd web
