+++ docs/pages/deploy/fly-io.mdx
@@ -96,7 +96,7 @@ Point the process at the config in `fly.toml`:
 
 ```toml
 [processes]
-  app = "/gestalt-server --config /etc/gestalt/config.yaml"
+  app = "/gestaltd --config /etc/gestalt/config.yaml"
 ```
 
 ## Deploy
