+++ docs/pages/deploy/railway.mdx
@@ -62,7 +62,7 @@ datastore:
 Override the container start command in Railway's service settings:
 
 ```
-/gestalt-server --config /build/toolshed/config.railway.yaml
+/gestaltd --config /build/toolshed/config.railway.yaml
 ```
 
 <Callout>
