+++ docs/pages/deploy/render.mdx
@@ -64,7 +64,7 @@ datastore:
 Set the **Docker Command** override:
 
 ```
-/gestalt-server --config config.render.yaml
+/gestaltd --config config.render.yaml
 ```
 
 ## Using SQLite instead
