+++ docs/pages/index.mdx
@@ -28,7 +28,7 @@ The upstream response passes through any configured **hooks** (response validato
 
 | Component | Description |
 |---|---|
-| **Go API server (`gestalt-server`)** | Configuration loading, auth provider management, token storage, operation dispatch, MCP endpoint, and webhook bindings. |
+| **Go API server (`gestaltd`)** | Configuration loading, auth provider management, token storage, operation dispatch, MCP endpoint, and webhook bindings. |
 | **Rust CLI (`gestalt`)** | Authenticate, list integrations, invoke operations, and manage API tokens from the terminal. |
 | **Next.js web UI** | Browse integrations, connect accounts, invoke operations interactively, and create API tokens. |
 | **Deployment assets** | Helm chart, Dockerfiles, and CI workflows for Kubernetes, Fly.io, Railway, Render, and AWS Lambda. |
