+++ docs/pages/tasks/run-locally.mdx
@@ -5,18 +5,10 @@ import { Callout } from 'nextra/components'
 ## Full stack (API + web UI)
 
 ```sh
-gestalt-server dev -config gestalt.dev.yaml
+gestaltd --config gestalt.dev.yaml
 ```
 
 This starts the API server and the Next.js web UI in a single process. The API runs on the port in your config (default 8080), the web UI on port 3000.
-
-## API only
-
-```sh
-gestalt-server dev -config gestalt.dev.yaml -api-only
-```
-
-Fastest loop for iterating on server behavior, config, or integration definitions.
 
 ## Web UI with mock API
 
@@ -26,12 +18,6 @@ cd gestalt/web
 ```
 
 Starts the Next.js app with built-in mock API routes. No Go server involved.
-
-## Override ports
-
-```sh
-gestalt-server dev -config gestalt.dev.yaml -api-port 9090 -web-port 4000
-```
 
 ## Development mode
 
@@ -58,7 +44,7 @@ Runs the API on port 8080 and the web UI on port 3000. Config is mounted from `d
 ## Validate your config
 
 ```sh
-gestalt-server config check -config gestalt.dev.yaml
+gestaltd --check --config gestalt.dev.yaml
 ```
 
 Prints resolved server settings, auth provider, datastore, integrations, runtimes, and bindings. Exits with "config ok" if valid.
