+++ docs/pages/tasks/use-with-mcp.mdx
@@ -48,7 +48,7 @@ Omit `providers` to expose all configured integrations.
 ## 3. Start the server
 
 ```sh
-gestalt-server dev -config config.yaml -api-only
+gestaltd --config config.yaml
 ```
 
 The MCP endpoint is available at `/mcp` on the same server as the REST API.
